Added Basic AI, Grid, and NavMesh2d
Added a Grid class that was originally going to help with pathfinding but realized there is an easier way but kept the class in since it could be useful for other stuff later. NavMesh 2d will help the pathfinding logic for the AI. Added basic AI that will stay idle until a target (Player) is in range. While in range the enemy will chase until it gets into attack range. Once in attack range it will stop and shoot the player. If the player leaves attack range the AI will go back to chasing.