Added Sounds PopupText and clean up
Added a bunch of sounds for picking up an item and the sound of guns. Will need to get more since these current ones do not fit and get annoying after a while. Also added popup text when the player is close to a buyable door and it shows the cost and what button to press to buy it.